+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## 0.0.2 (2018-08-?)
 
+### feat
+
+ - **navigation:** create navigation component
+
 ### docs
 
  - add changelog